requirements.txt
torch>=2.0.1
torchvision
torchaudio
aniso8601
numpy
attrs
click
Flask
flask-restx
flask_cors
itsdangerous
Jinja2
jsonschema
MarkupSafe
pyrsistent
pytz
six
Werkzeug
requests
Pillow
argparse
ultralytics
pytesseract
gunicorn